Define a separate UpdateSettings method
When Settings have changed, we need to call a different API endpoint.
This provides a separate Update method called UpdateSettings that
provides the new API endpoint and the slightly different payload.

This probably isn't ideal, and we may want Account Settings to be its
own Terraform resource. But this works for now as an iteration.