Ignore 409 status conflict on Workspace create
As part of retrying Create to work around 503 service unavailable
errors, we wanted to check for a 409 status conflict response because it
meant we didn't need to keep retrying the Create request.

However, when a 409 status conflict is reached, the Workspace is not
included in the response body. That means we'd have to run a separate
request to get the Workspace from the API - but the existing Get method
wants a workspace ID, which we don't have.

So we'd need to define a new GetByName method to get the method by the
name we originally wanted to use to create the workspace. But that
raises some questions:

- How do we handle a failure in that case? Do we add a retry at that
  point to avoid retrying the entire Create method and starting from
  square 1?
- Does it make sense to have a separate method just for this one error
  case?
- Would we ever even hit a 409 in the first place if we are sequentially
  running Create until the response is 201 created?
- We don't check for a 409 status conflict error on any other endpoints,
  is that a pattern we want to start following?

As an iteration, this commit simply removes the check for a 409 status
conflict error and relies on a simple retry mechanism. It will retry the
Create call until it returns 201 created.