Account for order differences in base_job_template
The base_job_template in the plan can differ from the base_job_template
in the work pool object retrieved from the API, specifically in the
order of the fields. In these cases, the objects are functionally
equivalent, but Terraform returns an error because it retrieved an
"inconsistent result after apply."

This change accounts for those differences by comparing the objects and
checking for equality using the go-cmp package. If the objects are not
equal, we will return an error diagnostic.