Use jsondiff package to return differences
The BaseJobTemplate field is large due to the number of required fields.
As a result, when there's a failure, printing the entire struct twice
(once for expected, once for got) is very hard to parse visually -
especially because Terraform logs them as JSON objects with quotes
escaped and with everything on one "line."

This package wraps go-cmp, which was used for comparison in the previous
commit. It adds on the ability to return the differences as a string,
which is much more readable in the output.

As a contrived example, when the resource retrieved from the API has
extra fields:

```
prefect_work_pool.example_work_pool Modifying...
╷
│ Error: Unexpected difference between base_job_template
│
│ Expected the provided base_job_template to be equal to the one retrieved from the API, differences: {
│     "prop-added":{"job_configuration": {}
│         "prop-added":{"auto_deregister_task_definition": "{{ auto_deregister_task_definition }}",}
│         "prop-added":{"cloudwatch_logs_options": "{{ cloudwatch_logs_options }}",}
```