Switch to go-test/deep for object comparison
This package allows comparing interfaces, so it's more flexible and
requires less conversion between types.

Since it's more flexible, also moves the function to a helper so we can
use it to compare other objects in the provider in the future.

Example of the output when the objects do not map:

```
=== RUN   TestAccResource_work_pool
    work_pool_test.go:58: Step 1/6 error: Check failed: Check 2/5 error: Found unexpected differences in work pool base_job_template: map[job_configuration]: map[command:{{ command }} env:{{ env }} labels:{{ labels }} name:{{ name }} stream_output:{{ stream_output }} working_dir:{{ working_dir }}] != <does not have key>
  ```