Block "data" argument: Clarify expected data format
Clarifies the expected data format for the "data" argument in the Block
resource.

The specific type is `schema.stringAttribute`, but we also set a
CustomType of `jsontypes.NormalizedType`, which is an RFC 7159 JSON
string.

The Terraform Documentation tooling does not consider the CustomType
field, so in the generated markdown it only say `String, Sensitive`.
This can be confusing for users who expect to just pass any string here.

So this change:
- Clarifies the format expectations in the `data` argument description
- Updates the examples to ensure we use `jsonencode` consistently

Related to https://github.com/PrefectHQ/terraform-provider-prefect/issues/255