Skip specify StorageDocumentID if unset
When a StorageDocumentID, which is of type UUID, is not provided, it
defaulted to a null UUID, which is `0000...`. This meant that if you
didn't provide a StorageDocumentID, you wouldn't be able to create a
deployment because it couldn't find a storage document by that null
UUID.

This uses pointers instead, so if it's not provided we get `nil` instead
of a "null" UUID (`0000...`).