Correctly test the name change
Before, when trying to test a variable resource name change, we were
actually creating an entirely new resource because we use the same
(randomized) value for the resource name and the attribute 'name'.

To ensure we're updating the same resource, this hard-codes the resource
name to 'test'. This shouldn't cause any conflicts with other tests
because the resource is created in an ephemeral workspace.