Extend implementation of ephemeral workspaces (#299)
* Create a Workspace struct

Uses a Workspace struct rather than returning multiple values, which
should be easier to use and extend if needed.

* Use new Workspace struct in tests

* Implement for datasources: variable, work pool, metadata

* Deployments: ephemeral workspace, implement a retry for block schemas

The block schemas are created asynchronously, and so may be empty
shortly after creation of a workspace. This is important in tests where
we make ephemeral workspaces.

* Flows: use ephemeral workspace, reuse GetResourceWorkspaceImportStateID

- Uses ephemeral workspace in flows test
- Reuses GetResourceWorkspaceImportStateID by moving it to the testing
  package since it's only used in tests

* Use new Workspace struct for deployment_access test

* Revert "Deployments: ephemeral workspace, implement a retry for block schemas"

This reverts commit 149ca01c56e1e4a81f5ff1994ddae02d671cdc7f.

* Revert block datasource ephemeral workspace

This one is facing the same race condition that Deployments are facing.
Will address in https://linear.app/prefect/issue/PLA-641/use-ephemeral-workspaces-in-resources-with-race-conditions