Use a pointer for 'active'
I was seeing an issue where my create payload would set active=false,
but the API would return a schedule where active=true. I noticed that
active=true is the default in the Prefect API, so I suspect something
about that conflict with Go's default value for active (false) has
something to do with it.

Changing this value to a pointer, and passing ValueBoolPointer in the
payloads and state configuration seems to fix the problem.

This definitely needs further thought.