Add helper functions for common HTTP requests
The HTTP requests across all of our PrefectClient-compatible clients are
very similar, leading to a lot of duplicate code. This leaves room for
error when adding new clients, and makes things a bit tougher to
maintain.

This change adds helper functions to replace this redundant code, coming
with all the usual benefits of DRY-ing things up.

This is also relevant as we aim for parity with the Prefect API, and the
less code we have to write, the better!

Related to https://linear.app/prefect/issue/PLA-690/cycle-6-catch-all