Deployments: support `schedule` (#304)
* Remove unused DeploymentFilter struct

* Implement schedules on Deployment resource

Initial implementation of schedules support by implementing it directly
in the deployment resource.

* Use go 1.22.7

* Implement schedule as separate resource

* Fix payload for Update

One schedule, not a list of them.

* Fix Update call: use scheduleID.String()

* Fix up schedule resource

- Copy schedule to model
- Validate schedule length of 1
- Add missing fields in payloads

* Test updates on a schedule

* Re-enable deployment update tests

* Make linting corrections

* Undo innocuous bracket move

* Add docs page

* Skip pointer values for fields

This was causing them to be added to the payloads even when they were
empty.

* Update tests for each kind of schedule

* Use a pointer for 'active'

I was seeing an issue where my create payload would set active=false,
but the API would return a schedule where active=true. I noticed that
active=true is the default in the Prefect API, so I suspect something
about that conflict with Go's default value for active (false) has
something to do with it.

Changing this value to a pointer, and passing ValueBoolPointer in the
payloads and state configuration seems to fix the problem.

This definitely needs further thought.

* Increase test coverage

* Add example usage for deployment_schedule

* Schedule ID: use state for unknown

This will use the state when the ID is not in the plan, which it won't
ever be because it's not configured by the user. This way we don't have
to get the current state just to pull the schedule ID.

* Mention that only one schedule is currently supported