Change ConcurrencyOptions to a custom type
ConcurrencyOptions was a types.Object, which made working with the
object more verbose for a resource that was a simple nested attribute.

Also, getting values from that object was causing the JSON request
payload to include escaped characters for some reason, meaning we needed
a helper function to strip them out.

This changes from types.Object to a custom Go type, allowing much easier
manipulation of the attribute - and removing the need for the custom
helper function to strip out quotes.