Only use optional for configurable attributes
From the docs on Optional:

Optional indicates whether the practitioner can choose to enter a value
for this attribute or not. Optional and Required cannot both be true.

Only `name` and `id` (and `{workspace,account}_id`) are configurable, so
ensure those are marked Optional and remove that field from ther est of
them.

This also matches the example from the docs on implementing datasources:
https://developer.hashicorp.com/terraform/tutorials/providers-plugin-framework/providers-plugin-framework-data-source-read