feat(deployments): add data source (#316)
* feat(deployments): add data source

Implements the data source for Deployments.

Closes https://linear.app/prefect/issue/PLA-799/add-datasource-support-for-deployments

Closes https://github.com/PrefectHQ/terraform-provider-prefect/issues/314

* Add the datasource to the provider

* Add tests for deployment datasource

* Add example, docs for deployment datasource

* Only use optional for configurable attributes

From the docs on Optional:

Optional indicates whether the practitioner can choose to enter a value
for this attribute or not. Optional and Required cannot both be true.

Only `name` and `id` (and `{workspace,account}_id`) are configurable, so
ensure those are marked Optional and remove that field from ther est of
them.

This also matches the example from the docs on implementing datasources:
https://developer.hashicorp.com/terraform/tutorials/providers-plugin-framework/providers-plugin-framework-data-source-read

* Generate Terraform Docs

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>