feat(deployments): support multiple schedules (#319)
* Add BaseModel struct for {re,data}sources

Copies the pattern from the `api` package to provide a reusable set of
fields common to many API objects.

* Support multiple schedules

The list schedules returned by Create should only include the schedule
associated with that resource, rather than all schedules on the
deployment, so we can use that to create the schedule and then persist
the correct one to the state.

Then for Read calls, we can get the correct schedule by ID.

* Add tests for multiple schedules

* Update the docs and examples for multiple schedules