feat(deployments): refactor pull step payload structure
Pull steps have a specific format in the API payload
that isn't very clearly defined by the API docs, since
the payload value is an empty list.

Looking more at the source code and documentation,
I adjusted the payload sent to the API to match the
expected format.

The format of the Terraform resource stays the same.

Overall changes:
* Introduced PullStepCommon to encapsulate shared fields across pull step types.
* Created specific types for PullStepGitClone,
  PullStepSetWorkingDirectory, and PullStepPullFrom to improve clarity and
  maintainability. This is where we can define the correct JSON field name
  for the payload.
* Updated the mapping functions to accommodate the new structure, ensuring proper handling of pull steps in both API and Terraform contexts.

Closes https://github.com/PrefectHQ/terraform-provider-prefect/issues/321