Actually test normalized value in resource
Previously, we were checking that the Deployment we retrieved had the
expected value for the base job template. This test goes one step
further by testing that the value we retrieved AND persisted to the
resource looks as expected. This means better coverage of our logic in
the provider, ensuring that we store the correct value in the resource.