Rename SafeUnmarshal to UnmarshalOptional
This gives a clearer name to the SafeUnmarshal function.
I originally thought we could remove this because I saw that an
attribute's Unmarshal function does check if it's null first, but the
real purpose of this function is to just be a helper to avoid having to
do a null check with IsNull on each optional attribute.

So this renames it to be more clear.