chore(JSON attributes): rely on Normalized for object equality (#340)
* Work pool: remove check for object equality

The JSON Normalized type mentions:

> Semantic equality logic is defined for Normalized such that
> inconsequential differences between JSON strings are ignored
> (whitespace, property order, etc).

Because of this, we should be able to rely on the normalization we do
with `NewNormalizedValue` to account for any differences between the
user-provided data and the data returned by the API.

* Actually test normalized value in resource

Previously, we were checking that the Deployment we retrieved had the
expected value for the base job template. This test goes one step
further by testing that the value we retrieved AND persisted to the
resource looks as expected. This means better coverage of our logic in
the provider, ensuring that we store the correct value in the resource.

* Deployments: test openapi schema attribute value

This tests the value of parameter_openapi_schema in the actual resource,
rather than just testing it in the received Deployment.

* Rename SafeUnmarshal to UnmarshalOptional

This gives a clearer name to the SafeUnmarshal function.
I originally thought we could remove this because I saw that an
attribute's Unmarshal function does check if it's null first, but the
real purpose of this function is to just be a helper to avoid having to
do a null check with IsNull on each optional attribute.

So this renames it to be more clear.

* Remove ObjectsEqual function

We no longer use this now that we're using jsontypes.Normalized
for comparison.

* Pass test pointer to NormalizedValueForJSON

This simplifies the error handling logic in the tests, since this helper
is only used there.

* Automation tests: use NormalizedValueForJSON

Instead of the custom TestCheckJSONAttr function, this change reuses the
new NormalizedValueForJSON test helper that provides the normalized JSON
value for expected values in tests.

* Complete usage of UnmarshalOptional

There were a few places where UnmarshalOptional (previously
SafeUnmarshal) wasn't implemented yet, so this fills that out.

* Move deployment JSON attribute marshaling to CopyDeploymentToModel

These calls were redundant, so moves them to the function for reuse.