Only set workspace handle filter if provided
Only sets the workspace handle filter if one is provided. Otherwise,
passes in an empty filter struct.

This allows the List method to return all workspaces. Before this
change, the empty "Any" field was causing no matches to be returned.

This will help us with the sweeper so it can return all workspaces and
let us filter the ones that match the prefix we use for acceptance
testing.