chore(tests): add sweeper to remove workspaces (#355)
* chore(tests): add sweeper to remove workspaces

Adds sweepers to clean up workspaces in cases where they were not
automatically cleaned up.

https://developer.hashicorp.com/terraform/plugin/testing/acceptance-tests/sweepers

Related to https://linear.app/prefect/issue/PLA-642/ensure-ephemeral-workspaces-are-cleaned-up-consistently

* Only set workspace handle filter if provided

Only sets the workspace handle filter if one is provided. Otherwise,
passes in an empty filter struct.

This allows the List method to return all workspaces. Before this
change, the empty "Any" field was causing no matches to be returned.

This will help us with the sweeper so it can return all workspaces and
let us filter the ones that match the prefix we use for acceptance
testing.

* Change sweeper to list all workspaces

The ephemeral workspaces we make during tests have random values.
Because sweepers are run in a separate command from the acceptance
tests, there's currently no way to persist those unique values between
commands so the sweeper can delete those specific workspaces.

Instead, as a first iteration, this change lists all workspaces in an
account and deletes any that match the acceptance testing prefix we use.
This is designed to run either by hand or by CI at a given interval,
such as overnight, when other acceptance tests are not running.

* Update Makefile to support sweepers

* Add sweep arg in Makefile doc

* Move sweeper to sweep package

Follows other patterns and allows us to test this specific package
instead of other packages with other tests.

* Split files into sweep.go, sweep_test.go

* Update testacc-dev to sweep ./internal/sweep

* Add CI workflow to sweep nightly

Adds a CI workflow that will run the Terraform sweepers each night.

---------

Co-authored-by: Edward Park <ed.sh.park@gmail.com>