chore: use default pull request template (#387)
* chore: use default pull request template

The support for multiple pull request templates does not expose an
option in the UI to select the template. This is only available by
providing query parameters to a URL for a new pull request specifying
the desired template.

That isn't quite what we want here - instead, we want the PR body to be
automatically filled in any time someone opens a PR. This change makes
use of the default PR template naming convention to make this happen.

Related to https://linear.app/prefect/issue/PLA-1114/implement-pr-hygiene-improvements

* Adjust template for default usage

* Mention labels

* Mention conventional commits

* Mention draft status