fix(accounts): fix the account domains payload
Fixes the payload type of account domains.

The API docs don't show the structure of the response, so we mistakenly
assumed it was the same as the request payload - a list of strings.

It's actually a list of objects, and the name is one of the fields in
those objects.

So this change accounts for that structure, and extracts the names from
it.

Closes https://github.com/PrefectHQ/terraform-provider-prefect/issues/390