Simplify 'skipFunc' logic
Rather than having another env var which you have to set to "yes",
just check to see if the user resource ID env var is set. If so, run the
tests.

This env var will be set by the script for user-only tests, and will be
set in CI for the user-only step.