chore(CI): fix user tests (#419)
* Pass 'TESTS' to 'make test'

* Mention value in 1P

* Support specific tests for 'test' target

* Test: inject failure to confirm tests running

* Use different invalid test

* Remove broken test setting

* Explain TESTS regex

* Add comments explaining API key source and 1P

* Add script to run user-only tests

* Simplify 'skipFunc' logic

Rather than having another env var which you have to set to "yes",
just check to see if the user resource ID env var is set. If so, run the
tests.

This env var will be set by the script for user-only tests, and will be
set in CI for the user-only step.