chore(docs): identify supported plans (#424)
* Install tfplugindocs, tomlv with mise

We can use mise experimental mode to install go binaries directly by URL
as they're not yet available in the mise registry.

* Call tfplugindocs via make, instead of go generate

The //go:generate approach was a bit less visible - we can actually call
the binary directly. This fits in better with our management of the
binary with mise.

* Add templates for all resources

* Implement templates for datasources

* Move example template to contributing docs

The template was being rendered into the 'docs/' dir.

* Remove basic/non-overridden templates

The basic templates add a lot of code to the repository.
We don't get much use out of them if we don't add additional notes
inside. So for now, let's remove them, and leave the documentation on
how to add a template if needed.

* Call tfplugindocs directly (via mise)

* Specify tiers for flows, accounts resources

* Specify pro and enterprise tiers for resources

* Rename description.go to plans.go

* Link to pricing page

* Specify plans for datasources

* Add comment on how to find plan support info

* Specify plans for webhook resource

* Add 'AllCloudPlans' var

* Specify OSS first

* Fill in missing plans for resources

* Consider plural/singular plans

* Workspace role: pro then enterprise

* Install tool dependencies with mise

* Run docs workflow changes on pull request

* Revert "Run docs workflow changes on pull request"

This reverts commit 9f0fbe283db055ba8cf0db6a60fcd14ad133374b.

Tested and job ran successfully