fix(deployments): deprecate `manifest_path` (#491)
* fix(deployments): deprecate `manifest_path`

Deprecates `manifest_path`. This is an old, unused field that Cloud only
has for backward compatibility for older clients but isn't required for
newer clients.

Related to https://linear.app/prefect/issue/PLA-1352/cycle-17-catch-all

* Keep manifest_path in the struct

Has to be there since it's in the schema.

Now that it's deprecated, we can remove Computed.