fix(work pools datasource): implement ID filter
The ID filter was present, but was not actually implemented in the
resource or client logic. This was causing failures in the OSS tests
because the nondeterministic order of work pools in the datasource could
lead to race conditions.

This uses the filter feature to get the expected work pool by ID, and
ensures the filter feature is properly implemented.

Related to https://linear.app/prefect/issue/PLA-191/run-acceptance-tests-against-prefect-server-in-docker-compose