fix(automations): support arrays in match_related (#537)
* fix(automations): support arrays in match_related

The match_related field in the API is an object _or_ an array of
objects. Using the `map[string]interface` type restricted the marshaling
to only work with objects and not arrays. Using the `any` type should
allow us to support both.

Closes https://github.com/PrefectHQ/terraform-provider-prefect/issues/536

* Unmarshal into any type

The shared helper function for unmarshaling optional fields
assumed `map[string]interface`, so this change skips that helper and
copies that logic with the desired `any` type.

If we end up doing this more we can add a helper for it.

Also adds a test for the scenario where `match_related` is an array
instead of an object.