docs(automations): list possible values for action type
Lists the possible values for action type.
These are already enumerated in a Go constant, so we reuse that in the
Markdown description for the field.

There are a couple open issues in the `terraform-plugin-docs` repository
to provide this functionality out of the box, given that the Validator
already ensures one of these values is used:

- https://github.com/hashicorp/terraform-plugin-docs/issues/243
- https://github.com/hashicorp/terraform-plugin-docs/issues/65