chore: replace Makefile with Mise tasks
The Makefile syntax can be tough to grok, and we aren't really
leveraging the main features of `make` in terms of compiling code
artifacts - we're using it more for providing easy access to short
commands.

`mise` has a tasks feature that offers this: https://mise.jdx.dev/tasks.
The syntax is much easier to read, and we get some nice benefits out of
the box like a help menu built from the `description` fields. We already
use `mise` to manage out tool versions, so this fits in nicely.

`mise` also has some nice features we can use in the future, including:
- Easy confirmation messages before running an important command
- Easy configuration for parallel jobs

This change replaces the Makefile with `mise` tasks. Open to input if we
think this is actually more usable/legible/maintainable.

Related to https://linear.app/prefect/issue/PLA-1624/cycle-24-catch-all