Put build artifact in hidden directory
Running a `mise` task returns an error when the build artifact
is visible. The only way I've found so far to "fix" this is to put
the build artifact in a hidden directory. Setting `task_disable_paths`
doesn't currently work because it always injects `mise-tasks` into
that path, which doesn't match where our binary is.

I suspect this is a bug but I haven't found any reports yet, and this
solution works well enough for now.